# Refactor lib/people_csv.rb