--- a/lib/people_csv.rb
+++ b/lib/people_csv.rb
@@ -77,45 +77,26 @@ class PeopleCSV < Array
   def guess_by_indexes(indexes)
     res = [output_header_row]
 
-    grouped_people = GroupedPeople.new(csv: self, indexes:)
+    grouped_peoples = GroupedPeoples.new(csv: self, indexes:)
 
-    grouped_people.each do |identifier_value, rows|
-      rows.each do |row|
-        res.append([identifier_value].concat(row))
+    grouped_peoples.each do |grouped_people|
+      grouped_people.people.each do |person|
+        res.append([grouped_people.identifier_value].concat(person))
       end
     end
 
     res
   end
 
-  # Private: Class for grouping people based on the given identifiers.
-  #          Key will be the value of identifier, and value will be an Array
-  #          object which contains the grouped people information.
+  # Private:
   #
-  # Example:
-  #   GroupedPeople.new(csv:, indexes:)
-  #   # => {
-  #          "take@example.com" => [
-  #            ["Take", "take@example.com", "09011111111"],
-  #            ["Takehiro", "take@example.com", "09022222222"]
-  #          ],
-  #          "09099999999" => [
-  #            ["Dave", "david@example.com", "09099999999"],
-  #            ["David", "dave@example.com", "09099999999"]
-  #          ]
-  #        }
-  #
-  class GroupedPeople < Hash
-    # csv - object of PeopleCSV
-    # indexes - indexes for the row in csv to group
-    def initialize(csv:, indexes:)
+  class GroupedPeoples < Array
+    def initialize(csv:, indexes:) # rubocop:disable Lint/MissingSuper
       indexes.each do |index|
         add_from_csv_by_index(csv:, index:)
       end
 
-      delete_non_grouped_data!
-
-      super
+      delete_non_grouped_people!
     end
 
     private
@@ -124,25 +105,43 @@ class PeopleCSV < Array
       csv.each_with_index do |row, row_index|
         next if row_index.zero?
 
-        value = row[index]
-        next if value.nil? # skip if the value of identifier is nil
-
-        # add new if there are no existing data
-        if self[value].nil?
-          self[value] = [row]
-        # add to existing data
-        else
-          self[value].append(row)
-        end
+        add_person(identifier_value: row[index], person: row)
       end
     end
 
-    # delete groups that doesn't have more than 1 row
-    def delete_non_grouped_data!
-      each do |identifier, rows|
-        delete(identifier) if rows.count < 2
+    def add_person(identifier_value:, person:)
+      return if identifier_value.nil?
+
+      grouped_people = find_by_identifier_value(identifier_value)
+      if grouped_people.nil?
+        append(GroupedPeople.new(identifier_value:, people: [person]))
+      else
+        grouped_people.people.append(person)
       end
     end
+
+    def find_by_identifier_value(identifier_value)
+      res = select { |grouped_people| grouped_people.identifier_value == identifier_value }
+      raise GroupedPeopleIdentifierValueNotUnique if res.count > 1
+
+      res.first
+    end
+
+    # delete groups that doesn't have more than 1 person
+    def delete_non_grouped_people!
+      delete_if { |grouped_people| grouped_people.people.count < 2 }
+    end
+
+    class GroupedPeople
+      attr_reader :identifier_value, :people
+
+      def initialize(identifier_value:, people:)
+        @identifier_value = identifier_value
+        @people = people
+      end
+    end
+
+    class GroupedPeopleIdentifierValueNotUnique < StandardError; end
   end
 
   class InvalidMatchingType < StandardError; end