* Change GroupedPeople class from Hash to Array and name it GroupedPeoples
* Create a new class GroupedPeople which will have the identifier value
  and people as attribute