@take 2022-07-10
## Performance check

### Before

```
❯ time ./bin/same_person_guesser.rb guess_by_matching_type \
    --input-file-destination './examples/inputs/input3.csv' \
    --matching_type 'same_phone' \
    --output-file-destination './examples/outputs/input3_grouped_by_same_phone.csv'
./bin/same_person_guesser.rb guess_by_matching_type --input-file-destination   0.25s user 0.09s system 89% cpu 0.376 total
```

### After

```
❯ time ./bin/same_person_guesser.rb guess_by_matching_type \
    --input-file-destination './examples/inputs/input3.csv' \
    --matching_type 'same_phone' \
    --output-file-destination './examples/outputs/input3_grouped_by_same_phone.csv'
./bin/same_person_guesser.rb guess_by_matching_type --input-file-destination   25.03s user 0.17s system 98% cpu 25.528 total
```